docx-to-json-ai/requirements.txt
python-dotenv
python-docx
pydantic
openai